[!!!][BUGFIX] Set permissions correctly
The change will allow graceful fallback to the default bucket
permissions for new objects to avoid users not being able to
operate on file ACLs via other tools (for example gsutil or
the cloud console).

It also removes the semantics of `private` and `publicRead`
CloudStorage API predefined ACLs in favour of default bucket
permissions and properly created additional ACLs on
visibility change.

Fixes: #3